Added rudimentary selection, Renamed Piece to PieceType, Other refactors
Added variables to ChessGame to handle position/piece selection
Added select() function to handle selection and initiate moves if appropriate. This function is not yet complete.
Added some debugging features to test selection.
Added selection logic to MainActivity.
Reworked GridLayout and TextViews to expand TextView 'hitbox' to entire grid square.
Encapsulated setting and getting of MainActivity.boardViewArray TextViews into functions.
Deprecated former move system and renamed ChessGame.move() to moveOld()
Encapsulated turn switching into private ChessGame function
Renamed ChessPiece.Piece enum to 'PieceType'. This was done to allow switching local variable names referencing this enum to be named 'type' instead of 'piece' which was ambiguous with 'ChessPiece' itself. This now allows 'piece' to be used for ChessPiece variables, and 'type' or 'pieceType' to be used to refer to the enum, overall clarifying and shortening code. Also replaced variables named 'chessPiece' with 'piece' and variables formerly named 'piece' with 'type' or 'pieceType'.
Changed ChessBoard.boardArray to be constant val, since it shouldn't be reassigned.
Changed ChessBoard.File.files to be called 'fileNames' to reflect that it's actually an array of strings, not an equivalent to File.values()
Changed default rank for Position to -1. Technically speaking, 0 might be 'more valid' than -1 since rank is zero-based in the board and grid layout arrays. I changed this to avoid any confusion.
Added equals() and hashCode() for ChessBoard.Position
Encapsulating castling and en passant logic into companion object functions that can be called from other classes, rather than forcing instantiation of a ChessMove to detect this (unnecessary, clunky, and I realized this would be annoying to deal with, if not impossible).
Changed ChessMove.capture to be of type ChessPiece rather than Piece(Type), to allow more information if needed.