Continued movement functionality, other minor changes
Started implementing ChessGame.move().
Added a (mutable) list of moves to ChessGame. This will be used to help keep track of game state, and it will also allow easy exporting to PGN later.
Updated documentation, especially for ChessPiece.
Reworked ChessMove to not take 'check' as an argument, as it was ambiguous. Its original meaning was `true` if the move led to a check (i.e., if a '+' would be added to the algebraic notation of the move). Instead, ChessMove will look at the game state to determine if the move leads to a check, and update game.inCheck appropriately. It will also use game.inCheck to determine if a player is currently in check and limit moves based on this.