Added dragging to move pieces
Added a log error for ChessBoard.File if an invalid index is passed for parsing/conversion.
Added an ID to the main layout for retrieval in program to add views to.
Added a semi-transparent black color for a shadow to show below the piece being dragged.
Added some documentation to ChessGame.
For ChessGame.select(), changed isPieceSelected to validate that the piece selected belongs to the player whose turn it is.
Renamed and reworked ChessGame selection and lastSelection into 'position' and 'lastPosition'. Also reworked select() to accept these new names and reduce ambiguity. select() also now returns isPieceSelected.
Added and initialized a few variables related to the followerView to MainActivity.
Refactored the creation of each TextView in boardGridLayout to the function createChessPieceView().
Replaced onClickListener for each TextView with an onTouchListener handling both 'click' functionality and 'drag' functionality.
Removed an initial call to performClick(). Not sure why that was there, it didn't seem to change anything.
Added the function createFollowerView() to initialize the follower view when a drag begins.
Added getPositionFromCoordinates() to transform a screen coordinate pair to a Position, for use with the drag feature.
Miscellaneous commenting/cleanup.