Started move validation
Added identifer for testing moves to ChessMove to accommodate cases where a move might be created that is not in sequence with the game state (as a prediction, for instance).
Added validation to ensure the King has not moved in ChessMove.castleValid().
Refactored and finished ChessMove.getEnPassantTarget().
Added documentation to ChessMove's companion object functions.
Added a private tracking variable for each king's position to ChessGame, and made sure to update it on a move.
Added an opponent function to the Player enum to easily retrieve the opposite color.
Implemented ChessGame.isCheckedAfterMove() and doLinearCheck() which together determine if the king is checked after a given move, to help with move validation.
Reworked ChessBoard.File.get() to return a nullable type, to ensure there is no ambiguity between a potentially viable move or position and an invalid one.
Added an optional argument to ChessBoard.set() to allow testing of various moves without updating the View, as in ChessGame.isCheckedAfterMove().
Added isOccupied() and getRelativePosition() to ChessBoard.
Started implementing getMovablePositions() and child functions in ChessPiece. Also refactored these to accept a Position as well as the instance of ChessGame.
Added a subclass PositionList to ChessPiece to allow easier addition of legal or valid moves.
Added getRankDirection to help with pawn move validation.
Implemented getKnightPositions and getPawnPositions.
Publicized the functions in ChessPiece's companion object so they can be used in ChessGame.isCheckedAfterMove().
Removed onClickListener from MainActivity (previously commented out and replaced by onTouchListener)