Finalized promotion
Added promotion string resources
Modified promotion dialog
Added logic to display correct color and orientation of pieces in promotion dialog
Changed `ChessMove.promotion` to a `Boolean`, and added a class member to hold the promoted piece. This is to allow working properly with the async nature of Android/dialogs.
Fixed an issue with piece positions not being removed from the `ChessBoard` piece position sets for each player upon captures.
Added a function to `ChessGame` for the controller (`MainActivity`) to call when a promotion result is determined from the dialog.
Encapsulated updating of `inCheck` into a function in `ChessGame`.
Encapsulated updating of actual check feedback (red gradient around king) into function in `MainActivity`.